pyproject.toml
[project]
name = "omnidesk-bot"
version = "0.1.0"
description = "Backend to manage Omnidesk tickets from a custom chat"
readme = "README.md"
authors = [
    { name = "FallenChromium", email = "fallenchromium@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115",
    "httpx>=0.27",
    "authlib>=1.3",
    "beanie>=1.26",
    "pydantic-settings>=2.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0",
]

[project.scripts]
omnidesk-bot = "omnidesk_bot:main"

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"
